Extract some testing utils.
